Add asignación automática de sector a tarea. Si pertenece se asgina la id, si no este valor es nulo

+++ Backend GestorTareas/src/main/java/com/tbd/GestorTareas/DTO/TareaRequestDTO.java
@@ -11,8 +11,8 @@ public class TareaRequestDTO {
     private Double latitud; // Para recibir la latitud
     private Double longitud; // Para recibir la longitud
     private boolean eliminado;
-    private Long usuario_id;
-    private Long sector_id;
+    private Integer usuario_id;
+    private Integer sector_id;
 
     // Getters y Setters
     public String getTitulo() { return titulo; }
@@ -39,11 +39,19 @@ public class TareaRequestDTO {
     public boolean isEliminado() { return eliminado; }
     public void setEliminado(boolean eliminado) { this.eliminado = eliminado; }
 
-    public Long getUsuario_id() { return usuario_id; }
-    public void setUsuario_id(Long usuario_id) { this.usuario_id = usuario_id; }
+    public Integer getUsuario_id() {
+        return usuario_id;
+    }
+    public void setUsuario_id(Integer usuario_id) {
+        this.usuario_id = usuario_id;
+    }
 
-    public Long getSector_id() { return sector_id; }
-    public void setSector_id(Long sector_id) { this.sector_id = sector_id; }
+    public Integer getSector_id() {
+        return sector_id;
+    }
+    public void setSector_id(Integer sector_id) {
+        this.sector_id = sector_id;
+    }
 
     // Método para generar el WKT desde latitud y longitud
     public String toWkt() {
